moved exceptions into utils
