+++ src/wxdata/cmc/gdps/url_scanner.py
@@ -7,7 +7,7 @@ This file hosts the URL Scanner for the GDPS
 import requests
 import sys
 import time
-import wxdata.cmc.gdps._exceptions as _exceptions
+import wxdata.cmc.utils._exceptions as _exceptions
 
 # Exception handling for Python >= 3.13 and Python < 3.13
 try:
