+++ src/wxdata/cmc/utils/_exceptions.py
